services/email-scheduler/database/db.py
from sqlalchemy.ext.asyncio import create_async_engine
import dotenv
import os
from database.tables import Base
import asyncio

dotenv.load_dotenv()

POSTGRES_USER = os.getenv("POSTGRES_USER")
POSTGRES_PASSWORD = os.getenv("POSTGRES_PASSWORD")
POSTGRES_DB = os.getenv("POSTGRES_DB")
POSTGRES_HOST = "postgresql"

engine = create_async_engine(
    f"postgresql+asyncpg://{POSTGRES_USER}:{POSTGRES_PASSWORD}@{POSTGRES_HOST}/{POSTGRES_DB}"
)

async def init_db():
    async with engine.begin() as conn:
        await conn.run_sync(Base.metadata.create_all)
    await engine.dispose()

async def get_db():
    async with engine.connect() as conn:
        yield conn

if __name__ == "__main__":
    asyncio.run(init_db())