services/email-scheduler/pyproject.toml
[project]
name = "email-scheduler"
version = "0.1.0"
description = ""
authors = [
    {name = "Roman Frolov",email = "r9888582703@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtplib (==5.1.0)",
    "amqp (==5.3.1)",
    "annotated-doc (==0.0.4)",
    "annotated-types (==0.7.0)",
    "anyio (==4.12.1)",
    "asyncpg (==0.31.0)",
    "billiard (==4.2.4)",
    "click (==8.3.1)",
    "click-didyoumean (==0.3.1)",
    "click-plugins (==1.1.1.2)",
    "click-repl (==0.3.0)",
    "colorama (==0.4.6)",
    "dnspython (==2.8.0)",
    "fastapi (==0.135.1)",
    "greenlet (==3.3.2)",
    "gunicorn (==25.1.0)",
    "h11 (==0.16.0)",
    "idna (==3.11)",
    "jinja2 (==3.1.6)",
    "kombu (==5.6.2)",
    "mako (==1.3.10)",
    "markupsafe (==3.0.3)",
    "packaging (==26.0)",
    "prompt-toolkit (==3.0.52)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "pydantic-core (==2.41.5)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.2.2)",
    "python-multipart (==0.0.22)",
    "six (==1.17.0)",
    "sqlalchemy (==2.0.48)",
    "starlette (==0.52.1)",
    "typing-inspection (==0.4.2)",
    "typing-extensions (==4.15.0)",
    "tzdata (==2025.3)",
    "tzlocal (==5.3.1)",
    "uvicorn (==0.42.0)",
    "vine (==5.1.0)",
    "wcwidth (==0.6.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "apscheduler (>=3.11.2,<4.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
